+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Fix missing an explicit requirement on `beberlei/assert` library
+- Add `expectedItemsCount` _optional_ argument to `parse` method of `ITuple` to allow explicit validation of parsed result
 
 ## 3.5.1 - 2018-07-13
 - Fix `implode` on `ListCollection`